pyproject.toml
[project]
name = "ai-intake"
version = "0.1.0"
description = "AI信息摄入系统 - 为工程师打造的低噪音AI信息订阅系统"
authors = [{name = "AI Intake Team"}]
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.11",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "rapidfuzz>=3.6.1",
    "PyYAML>=6.0.1",
    "openai>=1.54.0",
    "anthropic>=0.39.0",
    "python-dateutil>=2.8.2",
    "lxml>=5.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
